+++ app/src/main/java/com/moroom/android/ui/navui/dashboard/DashboardViewModel.java
@@ -1,14 +1,14 @@
-package com.moroom.android.navui.home;
+package com.moroom.android.ui.navui.dashboard;
 
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class DashboardViewModel extends ViewModel {
 
 //    private final MutableLiveData<String> mText;
 //
-//    public HomeViewModel() {
+//    public DashboardViewModel() {
 //        mText = new MutableLiveData<>();
-//        mText.setValue("This is home fragment");
+//        mText.setValue("This is dashboard fragment");
 //    }
 //
 //    public LiveData<String> getText() {
