+++ app/src/main/java/com/moroom/android/ui/navui/home/HomeViewModel.java
@@ -1,14 +1,14 @@
-package com.moroom.android.navui.profile;
+package com.moroom.android.ui.navui.home;
 
 import androidx.lifecycle.ViewModel;
 
-public class ProfileViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
 //    private final MutableLiveData<String> mText;
 //
-//    public ProfileViewModel() {
+//    public HomeViewModel() {
 //        mText = new MutableLiveData<>();
-//        mText.setValue("This is notifications fragment");
+//        mText.setValue("This is home fragment");
 //    }
 //
 //    public LiveData<String> getText() {
