+++ app/src/main/java/com/moroom/android/ui/navui/profile/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.moroom.android.navui.profile;
+package com.moroom.android.ui.navui.profile;
 
 import android.app.AlertDialog;
 import android.content.Intent;
