+++ app/src/main/java/com/moroom/android/ui/navui/profile/ProfileViewModel.java
@@ -1,14 +1,14 @@
-package com.moroom.android.navui.dashboard;
+package com.moroom.android.ui.navui.profile;
 
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class ProfileViewModel extends ViewModel {
 
 //    private final MutableLiveData<String> mText;
 //
-//    public DashboardViewModel() {
+//    public ProfileViewModel() {
 //        mText = new MutableLiveData<>();
-//        mText.setValue("This is dashboard fragment");
+//        mText.setValue("This is notifications fragment");
 //    }
 //
 //    public LiveData<String> getText() {
