+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.moroom.android.navui.home.HomeFragment"
+        android:name="com.moroom.android.ui.navui.home.HomeFragment"
         android:label="@string/home_fragment_label"
         tools:layout="@layout/fragment_home" >
 
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.moroom.android.navui.dashboard.DashboardFragment"
+        android:name="com.moroom.android.ui.navui.dashboard.DashboardFragment"
         android:label="@string/dash_fragment_label"
         tools:layout="@layout/fragment_dashboard" >
         <action
@@ -34,7 +34,7 @@
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.moroom.android.navui.profile.ProfileFragment"
+        android:name="com.moroom.android.ui.navui.profile.ProfileFragment"
         android:label="@string/profile_fragment_label"
         tools:layout="@layout/fragment_profile">
